Add fromObject, refactor model and page variables
